uapi|msft-internal: define IPE audit types as existing audit types
IPE audit events are out-of-tree definitions with new audit values.
The combination of these two facts results in an implicit dependency
on these values being valid in future releases of a kernel, which
is simply incorrect and misinformed.

To prevent kusto queries, and other future queries from breaking if a
future change takes the defines that IPE has currently defines, defining
IPE audit events in terms of existing audit events will take a hard
dependency on something set in stone.

Signed-off-by: Deven Bowers <deven.desai@linux.microsoft.com>
Reviewed-by: Tyler Hicks <tyhicks@linux.microsoft.com>

Link: https://microsoft.visualstudio.com/OS/_workitems/edit/28605821
TAG_MSFT: overlake