firmware: Introduce Persistent Logging feature to expose firmware boot logs
Firmware is publishing the current boot logs in DDR memory.
sochwmon is collecting the firmware boot logs for the previous boots,
but not for the current boot due to lack of /dev/mem in the OS.

The following patch exposes a sysfs entry(/sys/firmware/msft_fwlog)
to investigate the failures in firmware boot logs.

Steps to test the patch:
Pass earlyelog=0x8f103000,0x01000 as boot params
(Address: 0x8f103000, size: 0x1000)

Link: https://microsoft.visualstudio.com/OS/_workitems/edit/26984163/
TAG_MSFT: overlake
Signed-off-by: Allen Pais <apais@linux.microsoft.com>