Revert "netfilter: x_tables: Update remaining dereference to RCU"
This reverts commit 443d6e86f821a165fae3fc3fc13086d27ac140b1.

This (and the following) patch basically re-implemented the RCU
mechanisms of patch 784544739a25. That patch was replaced because of the
performance problems that it created when replacing tables. Now, we have
the same issue: the call to synchronize_rcu() makes replacing tables
slower by as much as an order of magnitude.

Revert these patches and fix the issue in a different way.

Signed-off-by: Mark Tomlinson <mark.tomlinson@alliedtelesis.co.nz>

Link: https://lore.kernel.org/r/20210308012413.14383-2-mark.tomlinson@alliedtelesis.co.nz
Signed-off-by: Tyler Hicks <tyhicks@linux.microsoft.com>